Add tests for loading sprites and textures with various caching strategies
