+++ Assets/_PackageRoot/Tests/Editor/TestLoading.Sprite.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d1c8a54a50b668844aa0ae2ee56fc731
+guid: 2a2187c670e0e4b4d867fcf6a84c229a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
