+++ Assets/_PackageRoot/Tests/Editor/TestLoading.Texture.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d1c8a54a50b668844aa0ae2ee56fc731
+guid: 4485204afe6ec1245aeb059479cc64e7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
